share common code for lock screen widget
